Refactor case detail page.
