+++ src/pages/caseDetail/caseDetailView.js
@@ -3,7 +3,6 @@ import {
   Grid,
   withStyles,
 } from '@material-ui/core';
-import { Link } from 'react-router-dom';
 import { CustomDataTable } from 'bento-components';
 import TableFooter from '@material-ui/core/TableFooter';
 import TableRow from '@material-ui/core/TableRow';
@@ -11,6 +10,8 @@ import TablePagination from '@material-ui/core/TablePagination';
 import StatsView from '../../components/Stats/StatsView';
 import { Typography } from '../../components/Wrappers/Wrappers';
 import icon from '../../assets/icons/Cases.Icon.svg';
+import formatBytes from '../../utils/formatBytes';
+import Subsection from '../../components/PropertySubsection/subsection';
 import CustomBreadcrumb from '../../components/Breadcrumb/BreadcrumbView';
 import {
   caseHeader,
@@ -18,17 +19,6 @@ import {
   rightPanelSubsections,
   tableConfig,
 } from '../../bento/caseDetailData';
-
-function formatBytes(bytes, decimals = 2) {
-  if (bytes === 0) return '0 Bytes';
-
-  const dm = decimals < 0 ? 0 : decimals;
-  const sizes = ['Bytes', 'KB', 'MB', 'GB', 'TB', 'PB', 'EB', 'ZB', 'YB'];
-
-  const i = Math.floor(Math.log(bytes) / Math.log(1024));
-
-  return `${parseFloat((bytes / (1024 ** i)).toFixed(dm))} ${sizes[i]}`;
-}
 
 const options = (classes) => ({
   selectableRows: 'none',
@@ -60,59 +50,6 @@ const options = (classes) => ({
     </TableFooter>
   ),
 });
-
-// Component to display a property
-const PropertyItem = ({
-  label, value, linkUrl, labelLinkUrl, classes,
-}) => {
-  const defaultValue = '';
-  return (
-    <Grid item container spacing={4}>
-      <Grid item xs={6}>
-        <span className={classes.title}>
-          { labelLinkUrl ? <Link to={labelLinkUrl.replace('{}', value)}>{label}</Link> : label }
-        </span>
-      </Grid>
-      <Grid item xs={6} className={classes.content}>
-        { value || value === 0 ? (
-          linkUrl ? <Link to={linkUrl.replace('{}', value)} className={classes.link}>{value}</Link>
-            : value
-        ) : defaultValue }
-      </Grid>
-    </Grid>
-  );
-};
-
-// Component to display a subsection
-const Subsection = ({ config, data, classes }) => (
-  <Grid item container className={classes.subsection}>
-    <Grid item container direction="column" className={classes.subsectionBody} xs={9}>
-      <Grid item>
-        <span className={classes.detailContainerHeader}>{config.sectionHeader}</span>
-      </Grid>
-      {
-        config.sectionDesc
-          ? (
-            <Grid item container className={classes.descriptionPart}>
-              <Grid item><span className={classes.description}>Description -</span></Grid>
-              <Grid item><span>{config.sectionDesc}</span></Grid>
-            </Grid>
-          ) : ''
-      }
-      {config.properties.map((prop) => (
-        <PropertyItem
-          key={prop.label}
-          label={prop.label}
-          value={data[prop.dataField]}
-          classes={classes}
-          linkUrl={prop.linkUrl}
-          labelLinkUrl={prop.labelLinkUrl}
-        />
-      ))}
-    </Grid>
-    <Grid item xs={3} />
-  </Grid>
-);
 
 // Main case detail component
 const CaseDetail = ({ data, classes }) => {
@@ -185,7 +122,7 @@ const CaseDetail = ({ data, classes }) => {
 
         <Grid container spacing={1} className={classes.detailContainer}>
           {/* Left panel */}
-          <Grid item sm={6} xs={12} className={classes.detailPannel}>
+          <Grid item sm={6} xs={12} className={classes.detailPanel}>
             <div className={classes.innerPanel}>
               <Grid container spacing={2}>
                 {leftPanelSubsections.map((section) => (
@@ -201,7 +138,7 @@ const CaseDetail = ({ data, classes }) => {
           </Grid>
           {/* Left panel end */}
           {/* Right panel */}
-          <Grid item sm={6} xs={12} className={classes.detailPannel}>
+          <Grid item sm={6} xs={12} className={classes.detailPanel}>
             <div className={classes.innerPanel}>
               <Grid container spacing={2}>
                 {rightPanelSubsections.map((section) => (
@@ -249,33 +186,12 @@ const CaseDetail = ({ data, classes }) => {
 };
 
 const styles = (theme) => ({
-  paddingLeft8: {
-    paddingLeft: '12px',
-  },
-  paddingBottm17: {
-    paddingBottm: '17px',
-  },
   container: {
     paddingTop: '38px',
     fontFamily: theme.custom.fontFamily,
     paddingLeft: '32px',
     paddingRight: '32px',
     background: '#FFFF',
-  },
-  content: {
-    fontSize: '12px',
-  },
-  marginTop23: {
-    marginTop: '24px',
-  },
-  warning: {
-    color: theme.palette.warning.main,
-  },
-  paper: {
-    textAlign: 'center',
-  },
-  fakeToolbar: {
-    ...theme.mixins.toolbar,
   },
   root: {
     fontFamily: theme.custom.fontFamily,
@@ -313,27 +229,6 @@ const styles = (theme) => ({
     fontWeight: 'bold',
     letterSpacing: '0.025em',
   },
-  headerSubTitleCate: {
-    color: '#606061',
-    fontWeight: 'bold',
-    fontFamily: theme.custom.fontFamilyRaleway,
-    textTransform: 'uppercase',
-    letterSpacing: '0.023em',
-    fontSize: '12px',
-    maxHeight: '30px',
-    overflow: 'hidden',
-    paddingLeft: '3px',
-  },
-  headerSubTitleContent: {
-    color: '#000000',
-    fontWeight: 'bold',
-    fontFamily: theme.custom.fontFamilyRaleway,
-    textTransform: 'uppercase',
-    letterSpacing: '0.023em',
-    fontSize: '12px',
-    paddingLeft: '3px',
-  },
-
   logo: {
     position: 'absolute',
     float: 'left',
@@ -350,19 +245,7 @@ const styles = (theme) => ({
     size: '12px',
     lineHeight: '23px',
   },
-  detailContainerHeader: {
-    textTransform: 'uppercase',
-    fontFamily: 'Lato',
-    fontSize: '17px',
-    letterSpacing: '0.025em',
-    color: '#0296C9',
-  },
-  detailContainerBottom: {
-    borderTop: '#81a6b9 1px solid',
-    marginTop: '8px',
-    padding: ' 35px 2px 63px 2px !important',
-  },
-  detailPannel: {
+  detailPanel: {
     paddingTop: '0 !important',
     paddingBottom: '0 !important',
     borderRight: '1px solid #81A6BA',
@@ -376,15 +259,6 @@ const styles = (theme) => ({
     paddingLeft: '16px',
     scrollbarColor: '#697270',
   },
-  subsectionBody: {
-    borderBottom: '1px solid #8DCAFF',
-    paddingBottom: '15px',
-  },
-  subsection: {
-    '&:last-child $subsectionBody': {
-      borderBottom: 'none',
-    },
-  },
   tableContainer: {
     background: '#f3f3f3',
   },
@@ -396,43 +270,6 @@ const styles = (theme) => ({
     margin: '0 auto auto auto',
     paddingTop: '50px',
   },
-  headerButtonLink: {
-    textDecoration: 'none',
-  },
-  button: {
-    borderRadius: '10px',
-    width: '178px',
-    height: '27px',
-    lineHeight: '18px',
-    fontSize: '10px',
-    color: '#ffffff',
-    textTransform: 'uppercase',
-    backgroundColor: '#ff8a00',
-    fontFamily: theme.custom.fontFamilySans,
-    '&:hover': {
-      backgroundColor: '#ff8a00',
-    },
-  },
-  detailContainerItems: {
-    paddingTop: '5px',
-    paddingLeft: '17px',
-    paddingRight: '16px',
-  },
-  title: {
-    color: '#9d9d9c',
-    fontFamily: theme.custom.fontFamilySans,
-    fontSize: '12px',
-    lineHeight: '12px',
-    letterSpacing: '0.017em',
-    fontWeight: '600',
-    textTransform: 'uppercase',
-  },
-  descriptionPart: {
-    paddingBottom: '26px',
-  },
-  description: {
-    fontWeight: 'bold',
-  },
   tableTitle: {
     textTransform: 'uppercase',
     fontFamily: 'Lato',
@@ -443,19 +280,6 @@ const styles = (theme) => ({
   },
   breadCrumb: {
     paddingTop: '3px',
-  },
-  paddingTop: {
-    paddingTop: '36px',
-  },
-  link: {
-    color: '#DD401C',
-    textDecoration: 'none',
-    '&:hover': {
-      textDecoration: 'underline',
-    },
-    '&:visited': {
-      color: '#9F3D26',
-    },
   },
   tableCell1: {
     paddingLeft: '25px',
